pyproject.toml
[project]
name = "final-project-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.91.0",
    "fastapi>=0.135.3",
    "litellm>=1.83.4",
    "mem0ai>=1.0.11",
    "pandas>=3.0.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.1",
    "strands-agents[a2a]>=1.34.1",
]
